@@ -11,28 +11,28 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^10.0.3",
-    "@angular/common": "^10.0.3",
-    "@angular/compiler": "^10.0.3",
-    "@angular/core": "^10.0.3",
-    "@angular/forms": "^10.0.3",
-    "@angular/platform-browser": "^10.0.3",
-    "@angular/platform-browser-dynamic": "^10.0.3",
-    "@angular/router": "^10.0.3",
+    "@angular/animations": "~11.0.1",
+    "@angular/common": "~11.0.1",
+    "@angular/compiler": "~11.0.1",
+    "@angular/core": "~11.0.1",
+    "@angular/forms": "~11.0.1",
+    "@angular/platform-browser": "~11.0.1",
+    "@angular/platform-browser-dynamic": "~11.0.1",
+    "@angular/router": "~11.0.1",
     "@auth0/angular-jwt": "^4.1.2",
     "@microsoft/signalr": "^5.0.1",
-    "@po-ui/ng-components": "^3.9.1",
-    "@totvs/po-theme": "^3.9.1",
+    "@po-ui/ng-components": "^4.3.0",
+    "@totvs/po-theme": "^4.3.0",
     "ngx-cookie-service": "^3.0.4",
     "rxjs": "~6.6.0",
     "tslib": "^2.0.0",
-    "zone.js": "~0.10.3"
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1000.3",
-    "@angular/cli": "^10.0.3",
-    "@angular/compiler-cli": "^10.0.3",
-    "@angular/language-service": "^10.0.3",
+    "@angular-devkit/build-angular": "~0.1100.1",
+    "@angular/cli": "~11.0.1",
+    "@angular/compiler-cli": "~11.0.1",
+    "@angular/language-service": "~11.0.1",
     "@types/node": "^12.11.1",
     "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
