@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnDestroy, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { Subscription } from 'rxjs';
 import { Pontuacao } from 'src/app/models/pontuacao.model';
@@ -12,9 +12,9 @@ import { SalaHubService } from 'src/app/services/sala-hub.service';
   selector: 'app-resultados',
   templateUrl: './resultados.component.html'
 })
-export class ResultadosComponent implements OnInit {
-  public maisVotado: string;
-  public pontuacao: Array<Pontuacao>;
+export class ResultadosComponent implements OnInit, OnDestroy {
+  public maisVotado = '';
+  public pontuacao: Array<Pontuacao> = [];
   public possoRevotarJogo: boolean;
   public possoConcluirJogo: boolean;
 
@@ -28,26 +28,33 @@ export class ResultadosComponent implements OnInit {
   private inscricaoReceberJogaodoes: Subscription;
   private inscricaoJogadorFinalizaJogo: Subscription;
   private inscricaoJogadorResetaJogo: Subscription;
+
   constructor(
     private activatedRoute: ActivatedRoute,
     private authService: AuthService,
     private salaHubService: SalaHubService
   ) {
-    this.sala = this.activatedRoute.snapshot.data['sala'];
+    const keySala = 'sala';
+    this.sala = this.activatedRoute.snapshot.data[keySala];
     this.salaConfig = this.sala.configuracao;
     this.administradores = this.sala.administradores;
     this.meuIdParticipante = authService.idParticipante;
     this.onReceberJogadores(this.sala.jogadores);
     this.possoRevotarJogo = salaHubService.possoResetarJogo;
     this.possoConcluirJogo = salaHubService.possoFinalizarJogo;
+
+    this.inscricaoSalaconfiguracao = this.salaHubService.receberConfiguracaoSala.subscribe(
+      (x: any) => this.onReceberConfiguracaoSala(x)
+      );
+    this.inscricaoReceberAdministradores = this.salaHubService.receberAdministradores.subscribe(
+      (x: Array<SalaParticipante>) => this.onReceberAdministradores(x)
+      );
+    this.inscricaoReceberJogaodoes = this.salaHubService.receberJogadores.subscribe((x: any) => this.onReceberJogadores(x));
+    this.inscricaoJogadorFinalizaJogo = this.salaHubService.jogadorFinalizaJogo.subscribe((x: any) => this.possoConcluirJogo = x);
+    this.inscricaoJogadorResetaJogo = this.salaHubService.jogadorResetaJogo.subscribe((x: any) => this.possoRevotarJogo = x);
   }
 
   ngOnInit(): void {
-    this.inscricaoSalaconfiguracao = this.salaHubService.receberConfiguracaoSala.subscribe(x => this.onReceberConfiguracaoSala(x));
-    this.inscricaoReceberAdministradores = this.salaHubService.receberAdministradores.subscribe(x => this.onReceberAdministradores(x));
-    this.inscricaoReceberJogaodoes = this.salaHubService.receberJogadores.subscribe(x => this.onReceberJogadores(x));
-    this.inscricaoJogadorFinalizaJogo = this.salaHubService.jogadorFinalizaJogo.subscribe(x => this.possoConcluirJogo = x);
-    this.inscricaoJogadorResetaJogo = this.salaHubService.jogadorResetaJogo.subscribe(x => this.possoRevotarJogo = x);
   }
 
   ngOnDestroy(): void {
@@ -56,7 +63,6 @@ export class ResultadosComponent implements OnInit {
     this.inscricaoReceberJogaodoes.unsubscribe();
     this.inscricaoJogadorFinalizaJogo.unsubscribe();
     this.inscricaoJogadorResetaJogo.unsubscribe();
-
   }
 
   private onReceberConfiguracaoSala(salaConfig: SalaConfiguracao): void {
@@ -71,7 +77,7 @@ export class ResultadosComponent implements OnInit {
     this.pontuacao = [];
 
     jogadores.forEach(jogador => {
-      let existeArray: boolean = false;
+      let existeArray = false;
 
       this.pontuacao.forEach(ponto => {
         if (ponto.cartaValor === jogador.votoCartaValor) {
@@ -85,10 +91,10 @@ export class ResultadosComponent implements OnInit {
       }
 
       this.pontuacao.sort((a: Pontuacao, b: Pontuacao) => {
-        let ret: number
+        let ret = 0;
 
         if (a.cartaValor === undefined) {
-          ret = 1
+          ret = 1;
         } else {
           ret = b.quantidadePontos - a.quantidadePontos;
           if (ret === 0) {
@@ -103,11 +109,11 @@ export class ResultadosComponent implements OnInit {
     });
   }
 
-  public revotarJogo() {
+  public revotarJogo(): void {
     this.salaHubService.enviarRevotarSala(this.sala.id);
   }
 
-  public concluirJogo() {
+  public concluirJogo(): void {
     this.salaHubService.enviarConcluirSala(this.sala.id);
   }
 
