@@ -6,21 +6,21 @@ import { Component, OnInit, Input, EventEmitter, Output } from '@angular/core';
   styleUrls: ['./carta.component.css']
 })
 export class CartaComponent implements OnInit {
-  @Output('p-click') click: EventEmitter<void> = new EventEmitter();
-  @Input("p-label") label: string;
-  @Input("p-selecionada") set selecionada (value: string) {
-    this.tipo = value == 'true' ? 'danger' : 'default';
+  @Output() pClick: EventEmitter<void> = new EventEmitter();
+  @Input() pLabel = '';
+  @Input('pSelecionada') set selecionada(value: boolean) {
+    this.tipo = value ? 'danger' : 'default';
   }
 
-  tipo: string = 'default';
+  tipo = 'default';
 
   constructor() { }
 
   ngOnInit(): void {
   }
 
-  cartaClick(){
-    this.click.emit();
+  cartaClick(): void {
+    this.pClick.emit();
   }
 
 }
