@@ -1,11 +1,11 @@
 import { Carta } from './carta.model';
 
 export class SalaConfiguracao {
-  public id: number;
-  public salaId: string;
-  public jogadorFinalizaJogo: boolean;
-  public jogadorResetaJogo: boolean;
-  public jogadorRemoveJogador: boolean;
-  public jogadorRemoveAdministrador: boolean;
-  public cartas: Array<Carta>;
+  public id?: number;
+  public salaId = '';
+  public jogadorFinalizaJogo = false;
+  public jogadorResetaJogo = false;
+  public jogadorRemoveJogador = false;
+  public jogadorRemoveAdministrador = false;
+  public cartas: Array<Carta> = [];
 }
