@@ -2,10 +2,10 @@ import { SalaConfiguracao } from './sala-configuracao.model';
 import { SalaParticipante } from './sala-participante.model';
 
 export class Sala {
-  public id: string;
-  public titulo: string;
-  public jogoFinalizado: boolean;
-  public configuracao: SalaConfiguracao;
-  public jogadores: Array<SalaParticipante>;
-  public administradores: Array<SalaParticipante>;
+  public id = '';
+  public titulo = '';
+  public jogoFinalizado = false;
+  public configuracao: SalaConfiguracao = new SalaConfiguracao();
+  public jogadores: Array<SalaParticipante> = [];
+  public administradores: Array<SalaParticipante> = [];
 }
