@@ -1,4 +1,5 @@
 export const environment = {
   production: true,
-  API: ""
+  API: 'http://localhost:5000/api/v1',
+  SOCKET: 'http://localhost:5000'
 };
