@@ -4,7 +4,8 @@
 
 export const environment = {
   production: false,
-  API: "http://localhost:5000/api/v1.0"
+  API: 'http://localhost:5000/api/v1',
+  SOCKET: 'http://localhost:5000'
 };
 
 /*
